Extract abstract model and admin classes
