+++ dvhb_hybrid/user_action_log/admin.py
@@ -1,30 +1,13 @@
 from django.contrib import admin
 
 from . import models
+from .base_admin import BaseUserActionLogEntryAdmin
 
 
 @admin.register(models.UserActionLogEntry)
-class UserActionLogEntryAdmin(admin.ModelAdmin):
-    list_display = (
-        'created_at', 'user', 'ip_address', 'message', 'type', 'subtype'
-    )
-    readonly_fields = (
-        'created_at', 'user', 'ip_address', 'message', 'type', 'subtype'
-    )
-    list_filter = [
-        'type',
-        'subtype',
-        ('created_at', admin.DateFieldListFilter)
-    ]
+class UserActionLogEntryAdmin(BaseUserActionLogEntryAdmin):
+    """
+    Concrete action log entry admin class to be used when utilizing hybrid's user_action_log app
+    """
 
-    def has_add_permission(self, request):
-        return False
-
-    def has_delete_permission(self, request, obj=None):
-        return False
-
-    def get_actions(self, request):
-        actions = super().get_actions(request)
-        if 'delete_selected' in actions:
-            del actions['delete_selected']
-        return actions
+    pass
